added revisions to eadheader, created namespace for ead revision
Extensive revisions including the removal of eadheader and frontmatter,
and addition of an EAC based control element. Patterns were copied from
EAC and in many cases revised (particularly substitution of a.common
and attrs-types for id, lang, and localType in EAC. Note that EAC
derived elements are in EAD namespace in this schema. EAD namespace for
this provisional revision is "http://plazi.org/ead/revised"